+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [UNRELEASED]
 * [#484] Replaced delegation list displayed under validator page with total number of delegations. 
+* [#509] Added display list of addresses available for user to login with when connecting the Ledger.  
 * [#497] Fetch keybase with timer settings 
 
 ## v0.41.x-12